Reorganize performance files
