+++ packages/devtools_app/lib/src/screens/performance/panes/timeline_events/legacy/event_details.dart
@@ -6,16 +6,16 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:vm_service/vm_service.dart' hide TimelineEvent;
 
-import '../../primitives/trace_event.dart';
-import '../../primitives/utils.dart';
-import '../../shared/common_widgets.dart';
-import '../../shared/globals.dart';
-import '../../shared/theme.dart';
-import '../profiler/cpu_profile_controller.dart';
-import '../profiler/cpu_profile_model.dart';
-import '../profiler/cpu_profiler.dart';
-import 'performance_controller.dart';
-import 'performance_model.dart';
+import '../../../../../primitives/trace_event.dart';
+import '../../../../../primitives/utils.dart';
+import '../../../../../shared/common_widgets.dart';
+import '../../../../../shared/globals.dart';
+import '../../../../../shared/theme.dart';
+import '../../../../profiler/cpu_profile_controller.dart';
+import '../../../../profiler/cpu_profile_model.dart';
+import '../../../../profiler/cpu_profiler.dart';
+import '../../../performance_controller.dart';
+import '../../../performance_model.dart';
 
 class EventDetails extends StatelessWidget {
   const EventDetails(this.selectedEvent);
