+++ packages/devtools_app/test/performance/timeline_flame_chart_test.dart
@@ -7,8 +7,8 @@ import 'dart:async';
 import 'package:devtools_app/src/charts/flame_chart.dart';
 import 'package:devtools_app/src/config_specific/ide_theme/ide_theme.dart';
 import 'package:devtools_app/src/config_specific/import_export/import_export.dart';
-import 'package:devtools_app/src/screens/performance/event_details.dart';
-import 'package:devtools_app/src/screens/performance/panes/timeline_events/timeline_flame_chart.dart';
+import 'package:devtools_app/src/screens/performance/panes/timeline_events/legacy/event_details.dart';
+import 'package:devtools_app/src/screens/performance/panes/timeline_events/legacy/timeline_flame_chart.dart';
 import 'package:devtools_app/src/screens/performance/performance_controller.dart';
 import 'package:devtools_app/src/screens/performance/performance_screen.dart';
 import 'package:devtools_app/src/screens/performance/tabbed_performance_view.dart';
